Extract Mosaic::Layout::AbstractMosaicLayout.
